requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-environ==0.10.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
multidict==6.0.4
python-slugify==8.0.1
python-telegram-bot==20.3
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
telegram-text==0.1.1
text-unidecode==1.3
urllib3==2.0.2
yarl==1.9.2
djangorestframework~=3.14.0
